fixed Demo solution
