--- demo/DemoApp.EFDataExample/DemoApp.EFDataExample.csproj
@@ -1,18 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <RootNamespace>DemoApp.EFDataExample</RootNamespace>
-    <AssemblyName>DemoApp.EFDataExample</AssemblyName>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.1" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\src\RulesEngine\RulesEngine.csproj" />
-  </ItemGroup>
-
-</Project>
