+++ demo/DemoApp/CancellationToken/Basic.cs
@@ -1,27 +1,29 @@
-﻿using RulesEngine.Models;
+﻿// Copyright (c) Microsoft Corporation.
+//  Licensed under the MIT License.
+
+using RulesEngine.Models;
 using System;
 using System.Collections.Generic;
-using System.Threading;
 using System.Threading.Tasks;
 using static RulesEngine.Extensions.ListofRuleResultTreeExtension;
 
-namespace DemoApp.Demos
+namespace DemoApp.CancellationToken
 {
-    public class CustomParameterName
+    public class Basic
     {
-        public async Task Run(CancellationToken ct = default)
+        public async Task Run(System.Threading.CancellationToken ct = default)
         {
             Console.WriteLine($"Running {nameof(Basic)}....");
 
             var workflows = new Workflow[] {
                 new Workflow {
-                    WorkflowName = "my_workflow",
-                    Rules =  new List<Rule>(){
+                    WorkflowName = "Test Workflow Rule 1",
+                    Rules = new List<Rule> {
                         new Rule {
-                            RuleName = "MatchesFabrikam",
-                            SuccessEvent = "does match",
-                            ErrorMessage = "does not match",
-                            Expression = "myValue.Value1 == \"Fabrikam\""
+                            RuleName = "Test Rule",
+                            SuccessEvent = "Count is within tolerance",
+                            ErrorMessage = "Over expected",
+                            Expression = "count < 3"
                         }
                     }
                 }
@@ -30,10 +32,10 @@ namespace DemoApp.Demos
             var bre = new RulesEngine.RulesEngine(workflows);
 
             var rp = new RuleParameter[] {
-                new RuleParameter("myValue", new { Value1 = "Fabrikam" })
+                new RuleParameter("input1", new { count = 1 })
             };
 
-            var ret = await bre.ExecuteAllRulesAsync("my_workflow", rp);
+            var ret = await bre.ExecuteAllRulesAsync("Test Workflow Rule 1", rp);
 
             var outcome = false;
