+++ demo/DemoApp/CancellationToken/BasicWithCustomTypes.cs
@@ -5,11 +5,10 @@ using RulesEngine.Models;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading;
 using System.Threading.Tasks;
 using static RulesEngine.Extensions.ListofRuleResultTreeExtension;
 
-namespace DemoApp.Demos
+namespace DemoApp.CancellationToken
 {
     public class BasicWithCustomTypes
     {
@@ -27,7 +26,7 @@ namespace DemoApp.Demos
             }
         }
 
-        public async Task Run(CancellationToken ct = default)
+        public async Task Run(System.Threading.CancellationToken ct = default)
         {
             Console.WriteLine($"Running {nameof(BasicWithCustomTypes)}....");
