+++ demo/DemoApp/CancellationToken/NestedInput.cs
@@ -8,10 +8,9 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
-using System.Threading;
 using System.Threading.Tasks;
 
-namespace DemoApp.Demos
+namespace DemoApp.CancellationToken
 {
     public class NestedInput
     {
@@ -21,12 +20,12 @@ namespace DemoApp.Demos
             public string Value { get; set; }
         }
 
-        public async Task Run(CancellationToken ct = default)
+        public async Task Run(System.Threading.CancellationToken ct = default)
         {
             Console.WriteLine($"Running {nameof(NestedInput)}....");
 
             var rp = new RuleParameter[] {
-                new RuleParameter("input1", new 
+                new RuleParameter("input1", new
                 {
                     SimpleProp = "simpleProp",
                     NestedProp = new {
@@ -52,7 +51,7 @@ namespace DemoApp.Demos
             var files = Directory.GetFiles(dir, "NestedInput.json", SearchOption.AllDirectories);
             if (files == null || files.Length == 0)
                 throw new Exception("Rules not found.");
-            
+
             var fileData = await File.ReadAllTextAsync(files[0], ct);
             var Workflows = JsonConvert.DeserializeObject<List<Workflow>>(fileData);
 
@@ -62,11 +61,9 @@ namespace DemoApp.Demos
             {
                 var ret = await bre.ExecuteAllRulesAsync(workflow.WorkflowName, rp);
 
-                ret.OnSuccess((eventName) =>
-                {
+                ret.OnSuccess((eventName) => {
                     Console.WriteLine($"evaluation resulted in success - {eventName}");
-                }).OnFail(() =>
-                {
+                }).OnFail(() => {
                     Console.WriteLine($"evaluation resulted in failure");
                 });
             }
